+++ roomfinder_website/model/roomModel.js
@@ -4,10 +4,17 @@ export default class RoomModel {
 
     // get room list route (room page)
     static getRoomList() {
+
+        const authToken = localStorage.getItem('authToken'); 
+        if (!authToken) {
+            throw new Error("You are not authorized");
+        }
+
         return fetch('http://localhost/RoomFinder_API/api/index.php/room', {
             method: 'GET',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${authToken}`
             }
         })
         .then(response => { 
@@ -26,10 +33,17 @@ export default class RoomModel {
 
     // get room by id route
     static getRoomById(roomId) {
+
+        const authToken = localStorage.getItem('authToken'); 
+        if (!authToken) {
+            throw new Error("You are not authorized");
+        }
+
         return fetch(`http://localhost/RoomFinder_API/api/index.php/room/${roomId}`, {
             method: 'GET',
             headers: {
                 'Content-Type': 'application/json',
+                'Authorization': `Bearer ${authToken}`
             }
         })
         .then(response => {
@@ -50,10 +64,17 @@ export default class RoomModel {
 
     static async createRoom(roomData) {
         try {
+
+            const authToken = localStorage.getItem('authToken'); 
+            if (!authToken) {
+                throw new Error("You are not authorized");
+            }
+
             const response = await fetch('http://localhost/RoomFinder_API/api/index.php/room', {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${authToken}`
                 },
                 body: JSON.stringify(roomData),
             });
@@ -61,29 +82,33 @@ export default class RoomModel {
             if (!response.ok) {
                 throw new Error(`HTTP error! Status: ${response.status}`);
             }
-
-            // Assuming the backend returns the room data directly on success
             const data = await response.json();
 
-            // Check if the returned data contains the expected room information
             if (data.room_number && data.room_building) {
-                return { success: true, room: data }; // Return the room data as the response
+                return { success: true, room: data }; 
             } else {
                 throw new Error('Invalid room data returned');
             }
         } catch (error) {
             console.error('Error creating room:', error.message);
-            throw error; // Rethrow the error for further handling in the ViewModel
+            throw error;
         }
     }
     
 
     // update room data route
     static updateRoom(roomId, updatedData) {
+
+        const authToken = localStorage.getItem('authToken'); 
+        if (!authToken) {
+            throw new Error("You are not authorized");
+        }
+
         return fetch(`http://localhost/RoomFinder_API/api/index.php/room/${roomId}`, {
             method: 'PATCH',
             headers: {
                 'Content-Type': 'application/json',
+                'Authorization': `Bearer ${authToken}`
             },
             body: JSON.stringify(updatedData),
         })
@@ -101,10 +126,17 @@ export default class RoomModel {
 
     // delete room route 
     static deleteRoomById(requestId) {
+
+        const authToken = localStorage.getItem('authToken'); 
+        if (!authToken) {
+            throw new Error("You are not authorized");
+        }
+        
         return fetch(`http://localhost/RoomFinder_API/api/index.php/room/${requestId}`, {
             method: 'DELETE',
             headers: {
                 'Content-Type': 'application/json',
+                'Authorization': `Bearer ${authToken}`
             },
         })
         .then(response => {
@@ -115,7 +147,6 @@ export default class RoomModel {
         })
         .then(data => {
             if (data.message === 'Room has been deleted') {
-                console.log('Room deleted:', data);
                 return { success: true, message: 'Room deleted successfully' };
             } else {
                 throw new Error('Failed to delete room');
